Refactor Configuration
Extracted the Configuration into its own module so that it's better
separated from the external interface. This will let me start to make
the configuration a bit more modular/updatable.
